+++ selenium_tests/test_admin_name_header_bug.py
@@ -1,12 +1,12 @@
-from selenium_tests.selenium_test_base import SeleniumTestBase
+from selenium_tests.AdminDriverTest import AdminDriverTest
 from selenium.webdriver.common.by import By
 
 
-class TestAdminNameHeaderBug(SeleniumTestBase):
+class TestAdminNameHeaderBug(AdminDriverTest):
     def test_admin_name_header_bug(self):
-        with self.logged_in("admin"):
+        with self.logged_in():
             self.click_element_located(By.LINK_TEXT, "Users")
 
-            self.click_button("Policies")
+            # self.click_button("Policies")
 
             self.wait_until_text_inside(By.CSS_SELECTOR, "span.hidden-xs", "admin")
