+++ selenium_tests/test_create_new_user.py
@@ -1,10 +1,10 @@
-from selenium_tests.selenium_test_base import SeleniumTestBase
+from selenium_tests.AdminDriverTest import AdminDriverTest
 from selenium.webdriver.common.by import By
 
 
-class TestCreateNewUser(SeleniumTestBase):
+class TestCreateNewUser(AdminDriverTest):
     def test_cancel(self):
-        with self.logged_in("admin"):
+        with self.logged_in():
             self.click_element_located(By.LINK_TEXT, "Users")
 
             self.click_new_entry_button()
@@ -12,7 +12,7 @@ class TestCreateNewUser(SeleniumTestBase):
             self.click_cancel_button()
 
     def test_create_user(self):
-        with self.logged_in("admin"):
+        with self.logged_in():
             self.click_element_located(By.LINK_TEXT, "Users")
 
             self.click_new_entry_button()
