+++ selenium_tests/test_login_logout.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-from selenium_tests.selenium_test_base import SeleniumTestBase
+from selenium_tests.RemoteAppDriverTest import RemoteAppDriverTest
 from selenium.webdriver.common.by import By
 
 
-class TestLoginLogout(SeleniumTestBase):
+class TestLoginLogout(RemoteAppDriverTest):
     def test_login_logout(self):
         self.login("test")
